refactor(metadata): extract 13 impl MetadataStore blocks into per-domain *_ops.rs
Split the single giant impl MetadataStore block in mod.rs into 13 per-domain
submodules, one per existing *_store.rs / *_sync.rs:

  profile_ops.rs          profile_sync delegations
  launcher_ops.rs         launcher_sync + launch_history delegations
  community_ops.rs        community_index + tap offline + trainer search
  collections_ops.rs      collections + favorites
  cache_ops.rs            cache_store delegations
  health_ops.rs           health_store delegations
  game_image_ops.rs       game_image_store delegations
  offline_ops.rs          offline_store + readiness + trainer hash
  version_ops.rs          version_store delegations
  config_history_ops.rs   config_history_store delegations
  preset_ops.rs           preset_store delegations
  catalog_ops.rs          optimization/readiness/proton catalog stores
  prefix_ops.rs           prefix_deps_store + prefix_storage_store

The inline SQL in index_community_tap_result (tap_id lookup + trainer sources
indexing) moved into a new community_index::index_community_tap_result_with_trainers
helper so the mod.rs wrapper becomes a one-line delegation.

No behavior change. No SQL change. Public API unchanged — every pub fn on
MetadataStore keeps the same path via multiple impl blocks on the same type.

mod.rs: 3632 → 2624 lines (−1008). The remaining ~2180 lines are the test
module, extracted in Phases 4–5.

Phase 2 of 6 for the metadata/mod.rs split.

Refs #291
Refs #290
