+++ src/crosshook-native/crates/crosshook-core/src/metadata/mod.rs
@@ -1,19 +1,31 @@
+mod cache_ops;
 mod cache_store;
+mod catalog_ops;
 mod collections;
+mod collections_ops;
 mod community_index;
+mod community_ops;
+mod config_history_ops;
 mod config_history_store;
 mod db;
+mod game_image_ops;
 mod game_image_store;
+mod health_ops;
 mod health_store;
 mod launch_history;
+mod launcher_ops;
 mod launcher_sync;
 mod migrations;
 mod models;
+mod offline_ops;
 pub(crate) mod offline_store;
 mod optimization_catalog_store;
 mod prefix_deps_store;
+mod prefix_ops;
 mod prefix_storage_store;
+mod preset_ops;
 mod preset_store;
+mod profile_ops;
 pub mod profile_sync;
 mod proton_catalog_store;
 mod readiness_catalog_store;
@@ -21,6 +33,7 @@ mod readiness_dismissal_store;
 mod readiness_snapshot_store;
 mod store;
 mod suggestion_store;
+mod version_ops;
 mod version_store;
 
 pub use game_image_store::GameImageCacheRow;
@@ -41,450 +54,12 @@ pub use readiness_snapshot_store::HostReadinessSnapshotRow;
 pub use store::MetadataStore;
 pub use version_store::{compute_correlation_status, hash_trainer_file};
 
-use crate::community::taps::CommunityTapSyncResult;
-use crate::discovery::TrainerSearchResponse;
-use crate::launch::diagnostics::models::DiagnosticReport;
-use crate::profile::{CollectionDefaultsSection, GameProfile, ProfileStore};
-use chrono::Utc;
-use rusqlite::{params, params_from_iter, OptionalExtension};
-use std::path::Path;
+use rusqlite::{params_from_iter, OptionalExtension};
 
 impl MetadataStore {
-    pub fn observe_profile_write(
-        &self,
-        name: &str,
-        profile: &GameProfile,
-        path: &Path,
-        source: SyncSource,
-        source_profile_id: Option<&str>,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("observe a profile write", |conn| {
-            profile_sync::observe_profile_write(
-                conn,
-                name,
-                profile,
-                path,
-                source,
-                source_profile_id,
-            )
-        })
-    }
-
-    pub fn lookup_profile_id(&self, name: &str) -> Result<Option<String>, MetadataStoreError> {
-        self.with_conn("look up a profile id", |conn| {
-            profile_sync::lookup_profile_id(conn, name)
-        })
-    }
-
-    pub fn query_profile_ids_for_names(
-        &self,
-        profile_names: &[String],
-    ) -> Result<Vec<(String, String)>, MetadataStoreError> {
-        if profile_names.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        self.with_conn("query profile ids for names", |conn| {
-            let placeholders = Self::in_clause_placeholders(profile_names.len());
-            let sql = format!(
-                "SELECT current_filename, profile_id \
-                 FROM profiles \
-                 WHERE deleted_at IS NULL AND current_filename IN ({placeholders})"
-            );
-            let mut stmt = conn
-                .prepare(&sql)
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "prepare query_profile_ids_for_names statement",
-                    source,
-                })?;
-            let rows = stmt
-                .query_map(params_from_iter(profile_names.iter()), |row| {
-                    Ok((row.get::<_, String>(0)?, row.get::<_, String>(1)?))
-                })
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "execute query_profile_ids_for_names",
-                    source,
-                })?;
-
-            let mut result = Vec::new();
-            for row in rows {
-                result.push(row.map_err(|source| MetadataStoreError::Database {
-                    action: "read a query_profile_ids_for_names row",
-                    source,
-                })?);
-            }
-            Ok(result)
-        })
-    }
-
-    pub fn observe_profile_rename(
-        &self,
-        old_name: &str,
-        new_name: &str,
-        old_path: &Path,
-        new_path: &Path,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("observe a profile rename", |conn| {
-            profile_sync::observe_profile_rename(conn, old_name, new_name, old_path, new_path)
-        })
-    }
-
-    pub fn observe_profile_delete(&self, name: &str) -> Result<(), MetadataStoreError> {
-        self.with_conn("observe a profile delete", |conn| {
-            profile_sync::observe_profile_delete(conn, name)
-        })
-    }
-
-    pub fn sync_profiles_from_store(
-        &self,
-        store: &ProfileStore,
-    ) -> Result<SyncReport, MetadataStoreError> {
-        self.with_conn("sync profiles from store", |conn| {
-            profile_sync::sync_profiles_from_store(conn, store)
-        })
-    }
-
-    pub fn observe_launcher_exported(
-        &self,
-        profile_name: Option<&str>,
-        slug: &str,
-        display_name: &str,
-        script_path: &str,
-        desktop_entry_path: &str,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("observe a launcher export", |conn| {
-            launcher_sync::observe_launcher_exported(
-                conn,
-                profile_name,
-                slug,
-                display_name,
-                script_path,
-                desktop_entry_path,
-            )
-        })
-    }
-
-    pub fn observe_launcher_deleted(&self, launcher_slug: &str) -> Result<(), MetadataStoreError> {
-        self.with_conn("observe a launcher deletion", |conn| {
-            launcher_sync::observe_launcher_deleted(conn, launcher_slug)
-        })
-    }
-
-    pub fn observe_launcher_renamed(
-        &self,
-        old_slug: &str,
-        new_slug: &str,
-        new_display_name: &str,
-        new_script_path: &str,
-        new_desktop_entry_path: &str,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("observe a launcher rename", |conn| {
-            launcher_sync::observe_launcher_renamed(
-                conn,
-                old_slug,
-                new_slug,
-                new_display_name,
-                new_script_path,
-                new_desktop_entry_path,
-            )
-        })
-    }
-
-    pub fn record_launch_started(
-        &self,
-        profile_name: Option<&str>,
-        method: &str,
-        log_path: Option<&str>,
-    ) -> Result<String, MetadataStoreError> {
-        self.with_conn("record a launch start", |conn| {
-            launch_history::record_launch_started(conn, profile_name, method, log_path)
-        })
-    }
-
-    pub fn record_launch_finished(
-        &self,
-        operation_id: &str,
-        exit_code: Option<i32>,
-        signal: Option<i32>,
-        report: &DiagnosticReport,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("record a launch finish", |conn| {
-            launch_history::record_launch_finished(conn, operation_id, exit_code, signal, report)
-        })
-    }
-
-    pub fn sweep_abandoned_operations(&self) -> Result<usize, MetadataStoreError> {
-        self.with_conn("sweep abandoned operations", |conn| {
-            launch_history::sweep_abandoned_operations(conn)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Phase 3: Community index
-    // -------------------------------------------------------------------------
-
-    pub fn index_community_tap_result(
-        &self,
-        result: &CommunityTapSyncResult,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("index a community tap", |conn| {
-            community_index::index_community_tap_result(conn, result)?;
-
-            // Index trainer sources for this tap after the profiles are committed.
-            if !result.index.trainer_sources.is_empty() {
-                let tap_url = &result.workspace.subscription.url;
-                let tap_branch = result
-                    .workspace
-                    .subscription
-                    .branch
-                    .as_deref()
-                    .unwrap_or("");
-                if let Some(tap_id) = conn
-                    .query_row(
-                        "SELECT tap_id FROM community_taps WHERE tap_url = ?1 AND tap_branch = ?2",
-                        params![tap_url, tap_branch],
-                        |row| row.get::<_, String>(0),
-                    )
-                    .optional()
-                    .map_err(|source| MetadataStoreError::Database {
-                        action: "look up tap_id for trainer sources indexing",
-                        source,
-                    })?
-                {
-                    community_index::index_trainer_sources(
-                        conn,
-                        &tap_id,
-                        &result.index.trainer_sources,
-                    )?;
-                }
-            }
-
-            Ok(())
-        })
-    }
-
-    pub fn lookup_community_tap_id(
-        &self,
-        tap_url: &str,
-        tap_branch: &str,
-    ) -> Result<Option<String>, MetadataStoreError> {
-        self.with_conn("look up community tap id", |conn| {
-            let mut stmt = conn
-                .prepare("SELECT tap_id FROM community_taps WHERE tap_url = ?1 AND tap_branch = ?2")
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "prepare community tap id lookup",
-                    source,
-                })?;
-            stmt.query_row(params![tap_url, tap_branch], |row| row.get(0))
-                .optional()
-                .map_err(|source| MetadataStoreError::Database {
-                    action: "look up community tap id",
-                    source,
-                })
-        })
-    }
-
-    /// Persists tap offline metadata after a successful sync (including cache-only sync).
-    /// No-op when the tap has not been indexed yet (no `community_taps` row).
-    pub fn upsert_community_tap_offline_from_sync_result(
-        &self,
-        result: &CommunityTapSyncResult,
-    ) -> Result<(), MetadataStoreError> {
-        let tap_url = &result.workspace.subscription.url;
-        let tap_branch = result
-            .workspace
-            .subscription
-            .branch
-            .as_deref()
-            .unwrap_or("");
-        let Some(tap_id) = self.lookup_community_tap_id(tap_url, tap_branch)? else {
-            return Ok(());
-        };
-        let now = Utc::now().to_rfc3339();
-        let size = crate::community::taps::directory_size_bytes(&result.workspace.local_path);
-        let last_sync_at = if result.from_cache {
-            self.lookup_community_tap_offline_state_row(&tap_id)?
-                .and_then(|r| r.last_sync_at)
-        } else {
-            Some(now)
-        };
-        self.upsert_community_tap_offline_state_row(
-            &tap_id,
-            1,
-            last_sync_at.as_deref(),
-            Some(size as i64),
-        )
-    }
-
-    pub fn list_community_tap_profiles(
-        &self,
-        tap_url: Option<&str>,
-    ) -> Result<Vec<CommunityProfileRow>, MetadataStoreError> {
-        self.with_conn("list community tap profiles", |conn| {
-            community_index::list_community_tap_profiles(conn, tap_url)
-        })
-    }
-
-    pub fn search_trainer_sources(
-        &self,
-        query: &str,
-        limit: i64,
-        offset: i64,
-    ) -> Result<TrainerSearchResponse, MetadataStoreError> {
-        self.with_conn("search trainer sources", |conn| {
-            crate::discovery::search_trainer_sources(conn, query, limit, offset)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Phase 3: Collections
-    // -------------------------------------------------------------------------
-
-    pub fn list_collections(&self) -> Result<Vec<CollectionRow>, MetadataStoreError> {
-        self.with_conn("list collections", |conn| {
-            collections::list_collections(conn)
-        })
-    }
-
-    pub fn create_collection(&self, name: &str) -> Result<String, MetadataStoreError> {
-        self.with_conn("create a collection", |conn| {
-            collections::create_collection(conn, name)
-        })
-    }
-
-    pub fn delete_collection(&self, collection_id: &str) -> Result<(), MetadataStoreError> {
-        self.with_conn("delete a collection", |conn| {
-            collections::delete_collection(conn, collection_id)
-        })
-    }
-
-    pub fn add_profile_to_collection(
-        &self,
-        collection_id: &str,
-        profile_name: &str,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("add a profile to a collection", |conn| {
-            collections::add_profile_to_collection(conn, collection_id, profile_name)
-        })
-    }
-
-    pub fn remove_profile_from_collection(
-        &self,
-        collection_id: &str,
-        profile_name: &str,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("remove a profile from a collection", |conn| {
-            collections::remove_profile_from_collection(conn, collection_id, profile_name)
-        })
-    }
-
-    pub fn list_profiles_in_collection(
-        &self,
-        collection_id: &str,
-    ) -> Result<Vec<String>, MetadataStoreError> {
-        self.with_conn("list profiles in a collection", |conn| {
-            collections::list_profiles_in_collection(conn, collection_id)
-        })
-    }
-
-    pub fn rename_collection(
-        &self,
-        collection_id: &str,
-        new_name: &str,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("rename a collection", |conn| {
-            collections::rename_collection(conn, collection_id, new_name)
-        })
-    }
-
-    pub fn update_collection_description(
-        &self,
-        collection_id: &str,
-        description: Option<&str>,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("update a collection description", |conn| {
-            collections::update_collection_description(conn, collection_id, description)
-        })
-    }
-
-    pub fn collections_for_profile(
-        &self,
-        profile_name: &str,
-    ) -> Result<Vec<CollectionRow>, MetadataStoreError> {
-        self.with_conn("list collections for a profile", |conn| {
-            collections::collections_for_profile(conn, profile_name)
-        })
-    }
-
-    pub fn get_collection_defaults(
-        &self,
-        collection_id: &str,
-    ) -> Result<Option<CollectionDefaultsSection>, MetadataStoreError> {
-        self.with_conn("read collection defaults", |conn| {
-            collections::get_collection_defaults(conn, collection_id)
-        })
-    }
-
-    pub fn set_collection_defaults(
-        &self,
-        collection_id: &str,
-        defaults: Option<&CollectionDefaultsSection>,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("write collection defaults", |conn| {
-            collections::set_collection_defaults(conn, collection_id, defaults)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Phase 3: Favorites
-    // -------------------------------------------------------------------------
-
-    pub fn set_profile_favorite(
-        &self,
-        profile_name: &str,
-        favorite: bool,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("set a profile favorite", |conn| {
-            collections::set_profile_favorite(conn, profile_name, favorite)
-        })
-    }
-
-    pub fn list_favorite_profiles(&self) -> Result<Vec<String>, MetadataStoreError> {
-        self.with_conn("list favorite profiles", |conn| {
-            collections::list_favorite_profiles(conn)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Phase 3: Cache
-    // -------------------------------------------------------------------------
-
-    pub fn get_cache_entry(&self, cache_key: &str) -> Result<Option<String>, MetadataStoreError> {
-        self.with_conn("get a cache entry", |conn| {
-            cache_store::get_cache_entry(conn, cache_key)
-        })
-    }
-
-    pub fn put_cache_entry(
-        &self,
-        source_url: &str,
-        cache_key: &str,
-        payload: &str,
-        expires_at: Option<&str>,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("put a cache entry", |conn| {
-            cache_store::put_cache_entry(conn, source_url, cache_key, payload, expires_at)
-        })
-    }
-
-    pub fn evict_expired_cache_entries(&self) -> Result<usize, MetadataStoreError> {
-        self.with_conn("evict expired cache entries", |conn| {
-            cache_store::evict_expired_cache_entries(conn)
-        })
-    }
-
     // -------------------------------------------------------------------------
     // Phase 3: Usage insights (inline SQL over launch_operations)
+    // Extracted in Phase 3 of the metadata/mod.rs split refactor. See #291.
     // -------------------------------------------------------------------------
 
     pub fn query_most_launched(
@@ -862,590 +437,6 @@ impl MetadataStore {
                 })?);
             }
             Ok(result)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Phase D: Health snapshot persistence
-    // -------------------------------------------------------------------------
-
-    pub fn upsert_health_snapshot(
-        &self,
-        profile_id: &str,
-        status: &str,
-        issue_count: usize,
-        checked_at: &str,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("upsert a health snapshot", |conn| {
-            health_store::upsert_health_snapshot(conn, profile_id, status, issue_count, checked_at)
-        })
-    }
-
-    pub fn load_health_snapshots(&self) -> Result<Vec<HealthSnapshotRow>, MetadataStoreError> {
-        self.with_conn("load health snapshots", |conn| {
-            health_store::load_health_snapshots(conn)
-        })
-    }
-
-    pub fn lookup_health_snapshot(
-        &self,
-        profile_id: &str,
-    ) -> Result<Option<HealthSnapshotRow>, MetadataStoreError> {
-        self.with_conn("look up a health snapshot", |conn| {
-            health_store::lookup_health_snapshot(conn, profile_id)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Game image cache persistence
-    // -------------------------------------------------------------------------
-
-    #[allow(clippy::too_many_arguments)]
-    pub fn upsert_game_image(
-        &self,
-        steam_app_id: &str,
-        image_type: &str,
-        source: &str,
-        file_path: &str,
-        file_size: Option<i64>,
-        content_hash: Option<&str>,
-        mime_type: Option<&str>,
-        source_url: Option<&str>,
-        expires_at: Option<&str>,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("upsert a game image cache entry", |conn| {
-            game_image_store::upsert_game_image(
-                conn,
-                steam_app_id,
-                image_type,
-                source,
-                file_path,
-                file_size,
-                content_hash,
-                mime_type,
-                source_url,
-                expires_at,
-            )
-        })
-    }
-
-    pub fn get_game_image(
-        &self,
-        steam_app_id: &str,
-        image_type: &str,
-    ) -> Result<Option<GameImageCacheRow>, MetadataStoreError> {
-        self.with_conn("get a game image cache entry", |conn| {
-            game_image_store::get_game_image(conn, steam_app_id, image_type)
-        })
-    }
-
-    pub fn evict_expired_images(&self) -> Result<Vec<String>, MetadataStoreError> {
-        self.with_conn_mut("evict expired game image cache entries", |conn| {
-            game_image_store::evict_expired_images(conn)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Offline trainer cache and readiness persistence
-    // -------------------------------------------------------------------------
-
-    pub fn check_offline_readiness_for_profile(
-        &self,
-        profile_name: &str,
-        profile_id: &str,
-        profile: &GameProfile,
-    ) -> Result<crate::offline::OfflineReadinessReport, MetadataStoreError> {
-        if !self.available {
-            return Err(MetadataStoreError::Corrupt(
-                "metadata store unavailable".to_string(),
-            ));
-        }
-        let Some(conn) = &self.conn else {
-            return Err(MetadataStoreError::Corrupt(
-                "metadata store connection missing".to_string(),
-            ));
-        };
-        let guard = conn.lock().map_err(|_| {
-            MetadataStoreError::Corrupt(
-                "metadata store mutex poisoned while check offline readiness".to_string(),
-            )
-        })?;
-        crate::offline::check_offline_preflight(profile_name, profile_id, profile, &guard)
-    }
-
-    pub fn upsert_trainer_hash_cache_row(
-        &self,
-        cache_id: &str,
-        profile_id: &str,
-        file_path: &str,
-        file_size: Option<i64>,
-        file_modified_at: Option<&str>,
-        sha256_hash: &str,
-        verified_at: &str,
-        created_at: &str,
-        updated_at: &str,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("upsert trainer hash cache", |conn| {
-            offline_store::upsert_trainer_hash_cache(
-                conn,
-                cache_id,
-                profile_id,
-                file_path,
-                file_size,
-                file_modified_at,
-                sha256_hash,
-                verified_at,
-                created_at,
-                updated_at,
-            )
-        })
-    }
-
-    pub fn lookup_trainer_hash_cache_row(
-        &self,
-        profile_id: &str,
-        file_path: &str,
-    ) -> Result<Option<TrainerHashCacheRow>, MetadataStoreError> {
-        self.with_conn("lookup trainer hash cache", |conn| {
-            offline_store::lookup_trainer_hash_cache(conn, profile_id, file_path)
-        })
-    }
-
-    pub fn verify_trainer_hash_for_profile_path(
-        &self,
-        profile_id: &str,
-        trainer_path: &Path,
-    ) -> Result<Option<crate::offline::HashVerifyResult>, MetadataStoreError> {
-        if !self.available {
-            return Ok(None);
-        }
-        let Some(conn) = &self.conn else {
-            return Ok(None);
-        };
-        let guard = conn.lock().map_err(|_| {
-            MetadataStoreError::Corrupt(
-                "metadata store mutex poisoned while verify trainer hash".to_string(),
-            )
-        })?;
-        crate::offline::verify_and_cache_trainer_hash(&guard, profile_id, trainer_path)
-    }
-
-    pub fn upsert_offline_readiness_snapshot_row(
-        &self,
-        profile_id: &str,
-        readiness_state: &str,
-        readiness_score: i64,
-        trainer_type: &str,
-        trainer_present: i64,
-        trainer_hash_valid: i64,
-        trainer_activated: i64,
-        proton_available: i64,
-        community_tap_cached: i64,
-        network_required: i64,
-        blocking_reasons: Option<&str>,
-        checked_at: &str,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("upsert offline readiness snapshot", |conn| {
-            offline_store::upsert_offline_readiness_snapshot(
-                conn,
-                profile_id,
-                readiness_state,
-                readiness_score,
-                trainer_type,
-                trainer_present,
-                trainer_hash_valid,
-                trainer_activated,
-                proton_available,
-                community_tap_cached,
-                network_required,
-                blocking_reasons,
-                checked_at,
-            )
-        })
-    }
-
-    pub fn load_offline_readiness_snapshot_rows(
-        &self,
-    ) -> Result<Vec<OfflineReadinessRow>, MetadataStoreError> {
-        self.with_conn("load offline readiness snapshots", |conn| {
-            offline_store::load_offline_readiness_snapshots(conn)
-        })
-    }
-
-    pub fn upsert_community_tap_offline_state_row(
-        &self,
-        tap_id: &str,
-        has_local_clone: i64,
-        last_sync_at: Option<&str>,
-        clone_size_bytes: Option<i64>,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn("upsert community tap offline state", |conn| {
-            offline_store::upsert_community_tap_offline_state(
-                conn,
-                tap_id,
-                has_local_clone,
-                last_sync_at,
-                clone_size_bytes,
-            )
-        })
-    }
-
-    pub fn lookup_community_tap_offline_state_row(
-        &self,
-        tap_id: &str,
-    ) -> Result<Option<CommunityTapOfflineRow>, MetadataStoreError> {
-        self.with_conn("lookup community tap offline state", |conn| {
-            offline_store::lookup_community_tap_offline_state(conn, tap_id)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Version snapshots
-    // -------------------------------------------------------------------------
-
-    pub fn upsert_version_snapshot(
-        &self,
-        profile_id: &str,
-        steam_app_id: &str,
-        steam_build_id: Option<&str>,
-        trainer_version: Option<&str>,
-        trainer_file_hash: Option<&str>,
-        human_game_ver: Option<&str>,
-        status: &str,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("upsert a version snapshot", |conn| {
-            version_store::upsert_version_snapshot(
-                conn,
-                profile_id,
-                steam_app_id,
-                steam_build_id,
-                trainer_version,
-                trainer_file_hash,
-                human_game_ver,
-                status,
-            )
-        })
-    }
-
-    pub fn lookup_latest_version_snapshot(
-        &self,
-        profile_id: &str,
-    ) -> Result<Option<VersionSnapshotRow>, MetadataStoreError> {
-        self.with_conn("look up the latest version snapshot", |conn| {
-            version_store::lookup_latest_version_snapshot(conn, profile_id)
-        })
-    }
-
-    pub fn load_version_snapshots_for_profiles(
-        &self,
-    ) -> Result<Vec<VersionSnapshotRow>, MetadataStoreError> {
-        self.with_conn("load version snapshots for profiles", |conn| {
-            version_store::load_version_snapshots_for_profiles(conn)
-        })
-    }
-
-    pub fn acknowledge_version_change(&self, profile_id: &str) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("acknowledge a version change", |conn| {
-            version_store::acknowledge_version_change(conn, profile_id)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Config revision history
-    // -------------------------------------------------------------------------
-
-    /// Insert a config revision for the profile, skipping if the hash matches the
-    /// latest recorded revision. Returns the new id when inserted, `None` on dedup.
-    pub fn insert_config_revision(
-        &self,
-        profile_id: &str,
-        profile_name_at_write: &str,
-        source: ConfigRevisionSource,
-        content_hash: &str,
-        snapshot_toml: &str,
-        source_revision_id: Option<i64>,
-    ) -> Result<Option<i64>, MetadataStoreError> {
-        self.with_conn_mut("insert a config revision", |conn| {
-            config_history_store::insert_config_revision(
-                conn,
-                profile_id,
-                profile_name_at_write,
-                source,
-                content_hash,
-                snapshot_toml,
-                source_revision_id,
-            )
-        })
-    }
-
-    /// List config revisions for a profile ordered newest first.
-    /// `limit` defaults to `MAX_CONFIG_REVISIONS_PER_PROFILE` when `None`.
-    pub fn list_config_revisions(
-        &self,
-        profile_id: &str,
-        limit: Option<usize>,
-    ) -> Result<Vec<ConfigRevisionRow>, MetadataStoreError> {
-        self.with_conn("list config revisions", |conn| {
-            config_history_store::list_config_revisions(conn, profile_id, limit)
-        })
-    }
-
-    /// Get a single config revision by id, scoped to `profile_id`.
-    /// Returns `None` if not found or the revision belongs to a different profile.
-    pub fn get_config_revision(
-        &self,
-        profile_id: &str,
-        revision_id: i64,
-    ) -> Result<Option<ConfigRevisionRow>, MetadataStoreError> {
-        self.with_conn("get a config revision", |conn| {
-            config_history_store::get_config_revision(conn, profile_id, revision_id)
-        })
-    }
-
-    /// Mark a revision as known-good for the profile, clearing the marker on all
-    /// other revisions for that profile (single-active-marker semantics).
-    pub fn set_known_good_revision(
-        &self,
-        profile_id: &str,
-        revision_id: i64,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("set a known-good config revision", |conn| {
-            config_history_store::set_known_good_revision(conn, profile_id, revision_id)
-        })
-    }
-
-    /// Clear the known-good marker from all revisions for the given profile.
-    pub fn clear_known_good_revision(&self, profile_id: &str) -> Result<(), MetadataStoreError> {
-        self.with_conn("clear known-good config revision markers", |conn| {
-            config_history_store::clear_known_good_revision(conn, profile_id)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Bundled / user launch optimization presets (metadata)
-    // -------------------------------------------------------------------------
-
-    pub fn list_bundled_optimization_presets(
-        &self,
-    ) -> Result<Vec<BundledOptimizationPresetRow>, MetadataStoreError> {
-        self.with_conn("list bundled optimization presets", |conn| {
-            preset_store::list_bundled_optimization_presets(conn)
-        })
-    }
-
-    pub fn get_bundled_optimization_preset(
-        &self,
-        preset_id: &str,
-    ) -> Result<Option<BundledOptimizationPresetRow>, MetadataStoreError> {
-        self.with_conn("get bundled optimization preset", |conn| {
-            preset_store::get_bundled_optimization_preset(conn, preset_id)
-        })
-    }
-
-    pub fn upsert_profile_launch_preset_metadata(
-        &self,
-        profile_id: &str,
-        preset_name: &str,
-        origin: ProfileLaunchPresetOrigin,
-        source_bundled_preset_id: Option<&str>,
-    ) -> Result<(), MetadataStoreError> {
-        let now = Utc::now().to_rfc3339();
-        self.with_conn("upsert profile launch preset metadata", |conn| {
-            preset_store::upsert_profile_launch_preset_metadata(
-                conn,
-                profile_id,
-                preset_name,
-                origin,
-                source_bundled_preset_id,
-                &now,
-            )
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Optimization catalog persistence
-    // -------------------------------------------------------------------------
-
-    pub fn persist_optimization_catalog(
-        &self,
-        entries: &[crate::launch::catalog::OptimizationEntry],
-        catalog_version: u32,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("persist optimization catalog", |conn| {
-            optimization_catalog_store::persist_optimization_catalog(conn, entries, catalog_version)
-        })
-    }
-
-    pub fn persist_readiness_catalog(
-        &self,
-        entries: &[crate::onboarding::HostToolEntry],
-        catalog_version: u32,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("persist readiness catalog", |conn| {
-            readiness_catalog_store::persist_readiness_catalog(conn, entries, catalog_version)
-        })
-    }
-
-    pub fn upsert_host_readiness_snapshot(
-        &self,
-        tool_checks: &[crate::onboarding::HostToolCheckResult],
-        detected_distro_family: &str,
-        all_passed: bool,
-        critical_failures: usize,
-        warnings: usize,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("upsert host readiness snapshot", |conn| {
-            readiness_snapshot_store::upsert_host_readiness_snapshot_impl(
-                conn,
-                tool_checks,
-                detected_distro_family,
-                all_passed,
-                critical_failures,
-                warnings,
-            )
-        })
-    }
-
-    pub fn get_host_readiness_snapshot(
-        &self,
-    ) -> Result<Option<HostReadinessSnapshotRow>, MetadataStoreError> {
-        self.with_conn("get host readiness snapshot", |conn| {
-            readiness_snapshot_store::get_host_readiness_snapshot_impl(conn)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Proton release catalog persistence (issue #274)
-    // -------------------------------------------------------------------------
-
-    /// Batch-upsert rows into `proton_release_catalog`.
-    pub fn put_proton_catalog(&self, rows: &[ProtonCatalogRow]) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("put proton catalog", |conn| {
-            proton_catalog_store::put_proton_catalog_impl(conn, rows)
-        })
-    }
-
-    /// Return all cached rows for `provider_id`, ordered by `fetched_at` descending.
-    pub fn get_proton_catalog(
-        &self,
-        provider_id: &str,
-    ) -> Result<Vec<ProtonCatalogRow>, MetadataStoreError> {
-        self.with_conn("get proton catalog", |conn| {
-            proton_catalog_store::get_proton_catalog_impl(conn, provider_id)
-        })
-    }
-
-    /// Delete all cached rows for `provider_id` (evict stale entries before a refresh).
-    pub fn clear_proton_catalog(&self, provider_id: &str) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("clear proton catalog", |conn| {
-            proton_catalog_store::clear_proton_catalog_impl(conn, provider_id)
-        })
-    }
-
-    /// Atomically replace all cached rows for a provider.
-    pub fn replace_proton_catalog(
-        &self,
-        provider_id: &str,
-        rows: &[ProtonCatalogRow],
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("replace proton catalog", |conn| {
-            proton_catalog_store::replace_proton_catalog_impl(conn, provider_id, rows)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Prefix dependency state persistence
-    // -------------------------------------------------------------------------
-
-    pub fn upsert_prefix_dep_state(
-        &self,
-        profile_id: &str,
-        package_name: &str,
-        prefix_path: &str,
-        state: &str,
-        error: Option<&str>,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("upsert prefix dep state", |conn| {
-            prefix_deps_store::upsert_dependency_state(
-                conn,
-                profile_id,
-                package_name,
-                prefix_path,
-                state,
-                error,
-            )
-        })
-    }
-
-    pub fn load_prefix_dep_states(
-        &self,
-        profile_id: &str,
-    ) -> Result<Vec<PrefixDependencyStateRow>, MetadataStoreError> {
-        self.with_conn("load prefix dep states", |conn| {
-            prefix_deps_store::load_dependency_states(conn, profile_id)
-        })
-    }
-
-    pub fn load_prefix_dep_state(
-        &self,
-        profile_id: &str,
-        package_name: &str,
-        prefix_path: &str,
-    ) -> Result<Option<PrefixDependencyStateRow>, MetadataStoreError> {
-        self.with_conn("load prefix dep state", |conn| {
-            prefix_deps_store::load_dependency_state(conn, profile_id, package_name, prefix_path)
-        })
-    }
-
-    pub fn clear_prefix_dep_states(&self, profile_id: &str) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("clear prefix dep states", |conn| {
-            prefix_deps_store::clear_dependency_states(conn, profile_id)
-        })
-    }
-
-    pub fn clear_stale_prefix_dep_states(&self, ttl_hours: i64) -> Result<u64, MetadataStoreError> {
-        self.with_conn_mut("clear stale prefix dep states", |conn| {
-            prefix_deps_store::clear_stale_states(conn, ttl_hours)
-        })
-    }
-
-    // -------------------------------------------------------------------------
-    // Prefix storage snapshot & cleanup audit persistence
-    // -------------------------------------------------------------------------
-
-    pub fn insert_prefix_storage_snapshot(
-        &self,
-        row: &PrefixStorageSnapshotRow,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("insert prefix storage snapshot", |conn| {
-            prefix_storage_store::insert_snapshot(conn, row)
-        })
-    }
-
-    pub fn list_latest_prefix_storage_snapshots(
-        &self,
-        limit: usize,
-    ) -> Result<Vec<PrefixStorageSnapshotRow>, MetadataStoreError> {
-        self.with_conn("list prefix storage snapshots", |conn| {
-            prefix_storage_store::list_latest_snapshots(conn, limit)
-        })
-    }
-
-    pub fn insert_prefix_storage_cleanup_audit(
-        &self,
-        row: &PrefixStorageCleanupAuditRow,
-    ) -> Result<(), MetadataStoreError> {
-        self.with_conn_mut("insert prefix storage cleanup audit", |conn| {
-            prefix_storage_store::insert_cleanup_audit(conn, row)
-        })
-    }
-
-    pub fn list_prefix_storage_cleanup_audit(
-        &self,
-        limit: usize,
-    ) -> Result<Vec<PrefixStorageCleanupAuditRow>, MetadataStoreError> {
-        self.with_conn("list prefix storage cleanup audit", |conn| {
-            prefix_storage_store::list_cleanup_audit(conn, limit)
         })
     }
 }
